refactoring testing structure in projects and implemented project updates - further project update webtests still in work
