+++ revolv/project/tests/test_tasks.py
@@ -1,0 +1,10 @@
+from django.test import TestCase
+from revolv.project.tasks import scrape
+
+class ScrapeTest(TestCase):
+    """Test that the scrape task runs with no errors,
+        and changes the project's solar data files"""
+
+    def test_scrape(self):
+        result = scrape.delay()
+        self.assertTrue(result.successful())
